refactor: renames UserInfo.ts and LoginInfo.ts
renames UserInfo.ts to UserState.ts and LoginInfo.ts to LoginState.ts
